+++ includes/builtins.h
@@ -6,7 +6,7 @@
 /*   By: minabe <minabe@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/11 13:54:35 by khorike           #+#    #+#             */
-/*   Updated: 2023/07/22 16:23:29 by minabe           ###   ########.fr       */
+/*   Updated: 2023/07/22 20:41:56 by minabe           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -100,7 +100,7 @@ void		handle_nodes(t_node *node, t_directory *dir,
 
 char		*search_quote(char **cmds);
 int			get_var_length(const char *str);
-void		change_normal_pluss(t_parse_context *ctx, t_parse_state *state);
+void		change_normal_plus(t_parse_context *ctx, t_parse_state *state);
 char		*quote_handle(char *str, t_directory *dir, t_env_var **env_vars);
 
 void	exec_pipe(t_node *node, t_directory *dir, t_env_var **env_vars);
