+++ srcs/builtins/ms_cd.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ms_cd.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: khorike <khorike@student.42.fr>            +#+  +:+       +#+        */
+/*   By: minabe <minabe@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/12 14:08:10 by khorike           #+#    #+#             */
-/*   Updated: 2023/07/20 14:54:45 by khorike          ###   ########.fr       */
+/*   Updated: 2023/07/22 20:51:05 by minabe           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
