+++ srcs/exec/builtins.c
@@ -6,11 +6,11 @@
 /*   By: minabe <minabe@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/22 15:39:21 by minabe            #+#    #+#             */
-/*   Updated: 2023/07/22 15:53:10 by minabe           ###   ########.fr       */
+/*   Updated: 2023/07/22 20:48:56 by minabe           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-# include "minishell.h"
+#include "minishell.h"
 
 bool	is_builtins(char *command)
 {
