+++ srcs/exec/select.c
@@ -6,15 +6,13 @@
 /*   By: minabe <minabe@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/13 17:26:36 by khorike           #+#    #+#             */
-/*   Updated: 2023/07/22 20:10:29 by minabe           ###   ########.fr       */
+/*   Updated: 2023/07/22 21:06:38 by minabe           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-#include "lexer.h"
-#include "token.h"
 
-static int	judgement_desuno(char **cmds, t_directory *dir, t_env_var **env_vars)
+int	judgement_desuno(char **cmds, t_directory *dir, t_env_var **env_vars)
 {
 	int	j;
 
@@ -36,50 +34,12 @@ static int	judgement_desuno(char **cmds, t_directory *dir, t_env_var **env_vars)
 	return (0);
 }
 
-static void	sub(char **cmds, t_directory *dir, t_env_var **env_vars)
+static int	sub(char **cmds, t_env_var **env_vars)
 {
 	if (ft_export(env_vars, cmds[1]))
 		exit(1);
 	else
-		dir->error = 0;
-	return ;
-}
-
-static void	support_fork(char **cmds)
-{
-	if (fork() == 0)
-	{
-		execve(cmds[0], cmds, NULL);
-		perror("execve failed");
-		exit(1);
-	}
-	else
-	{
-		wait(NULL);
-		return ;
-	}
-}
-
-static void	expansion(char **cmds, t_directory *dir)
-{
-	struct stat	s;
-
-	if (stat(cmds[0], &s) == 0)
-	{
-		if (S_ISDIR(s.st_mode))
-		{
-			printf("%s: is a directory\n", cmds[0]);
-			dir->error = 126;
-			return ;
-		}
-		else if (access(cmds[0], X_OK) == 0)
-		{
-			support_fork(cmds);
-		}
-	}
-	else
-		dir->error = execute_command(cmds[0], cmds) * 127;
-	return ;
+		return (0);
 }
 
 int	exec_builtin(char **cmds, t_directory *dir, t_env_var **env_vars)
@@ -87,27 +47,22 @@ int	exec_builtin(char **cmds, t_directory *dir, t_env_var **env_vars)
 	int			i;
 
 	i = 0;
-	if (judgement_desuno(cmds, dir, env_vars) == 1)
-		return ;
 	if (!ft_strcmp(cmds[0], "pwd"))
-		dir->error = ft_pwd(dir);
+		return (ft_pwd(dir));
 	if (!ft_strcmp(cmds[0], "cd"))
-		dir->error = ft_cd(dir, cmds[1], env_vars);
+		return (ft_cd(dir, cmds[1], env_vars));
 	if (!ft_strcmp(cmds[0], "exit"))
 		ft_exit();
 	if (!ft_strcmp(cmds[0], "env"))
-		dir->error = ft_env(*env_vars);
+		return (ft_env(*env_vars));
 	if (!ft_strcmp(cmds[0], "export"))
-		return (sub(cmds, dir, env_vars));
+		return (sub(cmds, env_vars));
 	if (!ft_strcmp(cmds[0], "unset"))
-		dir->error = ft_unset(env_vars, cmds[1]);
+		return (ft_unset(env_vars, cmds[1]));
 	while (cmds[i])
 		i++;
-	if (!ft_strcmp(cmds[0], "echo"))
-		dir->error = ft_echo(cmds, i - 1);
-	else if (ft_strcmp(cmds[0], "cd") && ft_strcmp(cmds[0], "pwd")
-		&& ft_strcmp(cmds[0], "unset"))
-		return (expansion(cmds, dir));
+	// if (!ft_strcmp(cmds[0], "echo"))
+	return (ft_echo(cmds, i - 1));
 }
 
 	// while (current)
