+++ srcs/expansion/helper_handle_quotation.c
@@ -6,7 +6,7 @@
 /*   By: minabe <minabe@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/21 17:02:15 by khorike           #+#    #+#             */
-/*   Updated: 2023/07/22 14:27:39 by minabe           ###   ########.fr       */
+/*   Updated: 2023/07/22 20:21:40 by minabe           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -57,7 +57,7 @@ int	get_var_length(const char *str)
 	return (len);
 }
 
-void	change_normal_pluss(t_parse_context *ctx, t_parse_state *state)
+void	change_normal_plus(t_parse_context *ctx, t_parse_state *state)
 {
 	*state = STATE_NORMAL;
 	(ctx->i)++;
