+++ srcs/minishell.c
@@ -6,7 +6,7 @@
 /*   By: minabe <minabe@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/11 14:16:03 by minabe            #+#    #+#             */
-/*   Updated: 2023/07/22 20:09:55 by minabe           ###   ########.fr       */
+/*   Updated: 2023/07/22 21:16:50 by minabe           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
